Added &lobby and &root. Changed the backend Entity code.
The root haver entity is now the House.
A House can contain rooms, users, services, and ghosts.
A Room can only contain users.

The room &lobby is a fake room which acts as a front-man for the House.
'house' is not a valid namespace, so you can not do 'INFO house ...'
There is no way of addressing the house, from the client-side, without using &lobby.
This is protocol-compatible with the previous system.

The &root user does nothing, except echo any private 'say' messages to the logfile.